Partition `Operation` enum
This splits `Operation` into three separate enums – `Control` (for
if/else, which get executed regardless of `vexec` and disabled
operations, which always fail, regardless of `vexec`), `Normal`, for
operations that respect `vexec`, and `Unknown` for undefined opcodes
which only fail if they’re on an active branch. This is done so that the
evaluator can be split on the same lines.

__NB__: I strongly recommend ignoring whitespace changes when reviewing
        this commit.

(cherry picked from commit f97b92c9b485ad1d3ca233749a1b5cd864ad800d)