Reduce redundant stack size checks
The `Stack` operations defined here already return `InvalidStackOperation` when they fail, so many
of the explicit `stack.is_empty()` checks are redundant.

However, removing them is delicate. Basically, they can only be removed if there is no other change
to the state that can occur before the `Stack` operation would  hit the same case. E.g., in
`OP_2DROP`, it can’t be removed because if there is one element on the stack, it’ll be removed
before the error is triggered, leaving the stack in different states between C++ and Rust.

Similarly, `OP_WITHIN` _can_ have the check removed, but the `pop`s can’t replace the `rget`s
because we might fail after the first `pop`, which would again leave us in a different state than
C++ which reads all the values before popping any of them.

Keeping the behavior identical to the C++ isn’t important per se, but it makes validating and
debugging changes easier, as we expect them to have the same state in every case.