Normalize errors more consistently
Well, my _tiny_ edge case of “only if the 202nd operation is a disabled opcode” didn’t slip past the
fuzzer. It caught that pretty quickly.

So, this does a better job of normalizing errors for comparisons.

First, it normalizes both the C++ and Rust side, which allows the Rust error cases to not be a
superset of the C++ error cases. Then, it also normalizes errors in the stepwise comparator (which I
think was done in #210, but it’s reasonable to do along with these other changes).

Also, `ScriptError::UnknownError` has been replaced with `ScriptError::ExternalError`, which takes
a string description. This is used for two purposes:

1. “Ambiguous” cases. One was already done – `UNKNOWN_ERROR` on the C++
   side with `ScriptError::ScriptNumError` or `ScriptError::HighS` on
   the Rust side, but now it’s handled differently. The other is the
   edge case I introduced earlier in this PR – Rust will fail with a
   `DisabledOpcode` and C++ will fail with a `OpCount`, so those two
   cases have been unified.
2. Errors that occur inside a stepper. Previously, this just melded in
   by returning `UnknownError`, but that was just the “least bad”
   option. Now we can distinguish these from other `ScriptError`.