Update the Rust toolchain
It was pinned to 1.81 because of a C++ linking issue in later versions. But now that 1.85 (and
later) has been released, we can move to “stable”, and bump the MSRV (because it was previously the
MaximumRSV, since later releases didn’t work).

This also makes bindgen track the MSRV more consistently for the code it generates.