Port script tests from C++
This is regex-transformed from the script_*valid.json tests in the C++
implementation.

The most significant difference is that this checks the specific result,
not just whether the script is valid. This is useful for a couple reasons:

- it catches places where a test may not be failing in the way we expect
 (e.g., invalid.rs has a test that is likely not testing what it should
   be);
- it catches places the implementation returns the wrong error (these
  have already been handled in the Rust implementation, but see all of
  the `UnknownError` results – if we’re explicitly testing for them,
  they shouldn’t be unknown, and one is definitely a case of forgot-to-
  report-`SigHighS`); and
- it allows us to interleave successful and unsuccessful tests in
  future, rather than splitting related tests into two files just
  because one succeeds and one represents an error.

However, some tests currently fail, this is because they rely on
specific transactions in `CHECK*SIG*`. This was easy to support prior to
the callback change, but these tests were never migrated to work on the
updated C++ implementation. See #240.