Update script tests
This brings them in sync with
https://github.com/bitcoin/bitcoin/commit/76da7613517d05aca31c2cee42b5ebf3d5c54974 ,
which is the last commit before they were merged into a single file.

Unfortunately, this also adds expected results to each entry, so the
diff is bigger than just added/changed tests.

Also, since the C++/JSON version of these tests was never updated to
work after the callback change, I’m updating the JSON files to match the
Bitcoin ones (modulo the tests we don’t support), just to keep
everything in sync.