Migrate script_*valid.json tests to Rust `TestVector`s (#241)
* Add more ergonomic opcode definitions

This was originally part of #223, but with the script validation tests being added, it was helpful
to bring this subset forward.

* Port script tests from C++

This is regex-transformed from the script_*valid.json tests in the C++
implementation.

The most significant difference is that this checks the specific result,
not just whether the script is valid. This is useful for a couple reasons:

- it catches places where a test may not be failing in the way we expect
 (e.g., invalid.rs has a test that is likely not testing what it should
   be);
- it catches places the implementation returns the wrong error (these
  have already been handled in the Rust implementation, but see all of
  the `UnknownError` results – if we’re explicitly testing for them,
  they shouldn’t be unknown, and one is definitely a case of forgot-to-
  report-`SigHighS`); and
- it allows us to interleave successful and unsuccessful tests in
  future, rather than splitting related tests into two files just
  because one succeeds and one represents an error.

However, some tests currently fail, this is because they rely on
specific transactions in `CHECK*SIG*`. This was easy to support prior to
the callback change, but these tests were never migrated to work on the
updated C++ implementation. See #240.

* Update script tests

This brings them in sync with
https://github.com/bitcoin/bitcoin/commit/76da7613517d05aca31c2cee42b5ebf3d5c54974 ,
which is the last commit before they were merged into a single file.

Unfortunately, this also adds expected results to each entry, so the
diff is bigger than just added/changed tests.

Also, since the C++/JSON version of these tests was never updated to
work after the callback change, I’m updating the JSON files to match the
Bitcoin ones (modulo the tests we don’t support), just to keep
everything in sync.

* Unify the valid & invalid script tests

This mirrors the changes from
https://github.com/bitcoin/bitcoin/commit/dde46d3ae1de18700e089bc1861cf1e18649dc5d

* Sync script tests with Bitcoin HEAD

This brings the script tests up to date as of
https://github.com/bitcoin/bitcoin/commit/fa33592898838cf92ae3bdb3ca9a522b5f9cce3d ,
but without the cases that require features we don’t support.

* Merge unified `TestVector`s into containing module