Respond to PR comments
The most significant change here is to how `HashType::from_bits` works.

Originally, it never failed if `VerificationFlags::StrictEnc` wasn’t
set, but it only ever passed around the byte. I had converted it to a
structured type, at which point it could fail if neither of the two
lowest bits were set, because there’s no defined interpretation for
`0b00`. But that’s a change in behavior, so this now stores the
`signed_outputs` field as an `Option`, with `None` representing the
`0b00` case. This is an invalid state, but has been permitted thus far.