Remove `enum_primitive` dependency
`enum_primitive` was last updated 8 years ago, and it depends on
`num-traits 0.1`. This causes a bunch of type confusion problems in
projects that depend on `num-traits 0.2` (as that crate used the
"compatibility trick" of re-exporting the `num-traits 0.2` API inside
the last release of `num-traits 0.1`, which today causes ambiguous trait
resolution errors).