Include opcode with all interpreter errors
Previously, some interpreter errors included the opcode, but others
didn’t. However, all interpreter errors occur when processing an
opcocde, so they should all include the relevant opcode.

This also fixes a few inconsistencies.

`interpreter::Error::UnbalancedConditional` was used for three purposes
– if the stack didn’t contain a conditional, if an `OP_ELSE` or
`OP_ENDIF` occurred without an `OP_(NOT)IF`, and if and `OP_ENDIF` was
missing.

The first of these is better represented by
`interpreter::Error::InvalidStackOperation`, and the last is really
about the entire script, not processing an opcode, so it’s now
represented by `script::Error::UnclosedConditional`. These all normalize
to the previous value, to keep consistency wtih C++.