Unify `interpreter::Error::Invalid*tackOperation`
No operation reads from both stacks, so now that we include the opcode
with the error, these can be the same type. (And if they do need to be
distinguished in future, it should be with a simple `Stack|Altstack`
enum[^1]).

[^1]: `interpreter::Error::UnbalancedConditional` is also (since
breaking it apart in a previous commit) the same error for a different
stack and can also be distinguished by opcode, but it would need the
enumeration (with an `Exec` variant) in order to maintain compatibility
with the C++ implementation.