Constrain the mutability of `State`
Take advantage of Rust’s affine type system to only require `mut` at the
level of functions that actually manipulate the state.

This gives us better separation from the C++ implementation – we no
longer need to keep the state in sync _during_ a step. If a step fails,
we no longer have a partially-mutated state that needs to be in sync
with partially-mutated state from the C++ side..