Replace an `assert!` with `expect`
We asserted, then handled the error covered by the assertion. That error
was also annoying, because it was the one case of an interpreter error
that wasn’t associated with a specific opcode[^1]. Now both the
assertion and error handling are covered by an `expect`, which
eliminates the no-opcode interpreter error case.

[^1]: I was going to change this error to be a different `script::Error`
case (which would be normalized back to the `interpreter::Error`), but
then realized it was guarded by the assert. But this would still be an
acceptable alternative.