Make `PushSize` an `opcode::Error`
It was previously an `interpreter::Error`. This change requires two
adjustments for C++ compatibility (commented), but it’s a win because
it’s one more thing that can be identified statically and it lets us
further restrict the upper bound on `LargeValue`s.

This also changes the relevant tests to use hex strings (`H`) instead of
text (`A`), because we require text to be a valid `PushValue`, which
those no longer are.