Reduce usage of `normalized_err` in test vectors
`normalized_err` discards some error information that the Rust side has
but the C++ side doesn’t. Sometimes it’s necessary because constructing
the exact error case manually can be difficult (e.g., you can’t directly
construct a `TryFromIntError`). However, in other cases, it was just
used for expediency.

This gets rid of it for the expedient cases, and introduces a few more
`From` instances for nested `enum`s for ergonomics.