Handle some `Option` values more explicitly
One of the `Option`s is in `get_sig_op_count`, where there are multiple
cases that result in the same outcome. This distinguishes those cases
and adds a comment to each.

The other is n signature decoding. In #259 the fuzzer caught a panic
that should have been `_checked`. Adding the check gave us an `Option`
that I initially ignored (other than the desired case). It now reports
different errors for the `Some` and `None` cases. This also makes the
`value` field of the `InvalidDerEncoding::InvalidComponent` variant
non-optional.