Add tests for the “modern” API
There are test vectors for a “purely authored” `Script`, but when
creating a transaction, we are likely testing an authored script sig
against a script pubkey from the chain. This ensures (as much as
possible) that that behavior is aligned.

This also adds a `script::FromChain` alias and a `Code::to_component()`
function to improve the ergonomics.